Moved AllpassFDNMatrix to functional
And re-organized the functional module for better readability
